+++ src/tool/cpp/basewinrt/base.h
@@ -30,21 +30,7 @@
 #include "base_chrono.h"
 #include "base_security.h"
 #include "base_await.h"
-#include "base_collections.h"
-#include "base_collections_base.h"
-#include "base_collections_input_iterable.h"
-#include "base_collections_input_vector_view.h"
-#include "base_collections_input_map_view.h"
-#include "base_collections_input_vector.h"
-#include "base_collections_input_map.h"
-#include "base_collections_vector.h"
-#include "base_collections_map.h"
 #include "base_std_hash.h"
 #include "base_std_fire_and_forget.h"
-#include "base_std_async_action.h"
-#include "base_std_async_action_with_progress.h"
-#include "base_std_async_operation.h"
-#include "base_std_async_operation_with_progress.h"
 #include "base_reflect.h"
 #include "base_natvis.h"
-
